% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{area_triangle}
\alias{area_triangle}
\title{Triangle}
\usage{
area_triangle(base, height)
}
\arguments{
\item{base}{numeric the base of the triangle}

\item{height}{numeric the height of the triangle}
}
\value{
numeric value of the area of the triangle
}
\description{
Triangle
}
